+++ packages/ui/src/components/flyout/components/flyout-backdrop/index.ts
@@ -1,0 +1,2 @@
+export { FlyoutBackdrop } from './flyout-backdrop.component.js';
+export { type FlyoutBackdropProps, type FlyoutBackdropRenderProps } from './flyout-backdrop.types.js';