+++ packages/ui/src/components/flyout/index.ts
@@ -1,0 +1,2 @@
+export { Flyout } from './flyout.component.js';
+export { type FlyoutProps } from './flyout.types.js';